+++ pkg/bpfassets/attacher/bcc_attacher_stub.go
