+++ pkg/bpfassets/attacher/bpf_perf.go
