+++ pkg/collector/metric/utils.go
@@ -17,7 +17,7 @@ limitations under the License.
 package metric
 
 import (
-	"github.com/sustainable-computing-io/kepler/pkg/attacher"
+	"github.com/sustainable-computing-io/kepler/pkg/bpfassets/attacher"
 	"github.com/sustainable-computing-io/kepler/pkg/config"
 	"github.com/sustainable-computing-io/kepler/pkg/model"
